Move auto_tests into MS05Utils
